added health manager, replaced health enemy and player health, added game object functions for relations, refactored skills and controllers

+++ Assets/Scripts/Enemy/MageMovement.cs
@@ -22,7 +22,7 @@ namespace LIL
         private Animator anim;
         private LightFuel torchLight;
         private bool moveCancelled = false;
-        private int currentDamage;
+        private float currentDamage;
 
         void Start()
         {
